remove HashMap in REST resource method
changed the Prefs resource to use a String instead of a HashMap, because
JAXB was giving me grief about serialising the HashMap. we currently
only have 1 preference anyway (the timezone id) so the HashMap was
overkill.

strangely enough this works in the master branch, probably because we're
using a different version of JAXB.